Added base source code
Added the source code files. There are 4 screens (Home, Legends, Weapons, Heirloom) in the app, each screen has its own source code file (the app is not coded in just 1 file). A bottom tab navigator is used to contain each of the 4 screens. The "Legends" screen uses stack navigation for each of the 20 legends, and the "Weapons" screen uses stack navigation to navigate to any of 12 screens (6 weapon types' screens + 6 weapon attachments' screens). 
The readme file needs to be updated to reflect some additions to the app.